--- src/main/java/com/jarqprog/artGallery/domain/exception/CreationException.java
@@ -1,10 +1,0 @@
-package com.jarqprog.artGallery.domain.exception;
-
-import com.jarqprog.artGallery.domain.dto.DTO;
-
-public class CreationException extends RuntimeException {
-
-    public CreationException(Class<? extends DTO> clazz, String message) {
-        super(String.format("Cannot create %s. Cause: %s", clazz.getSimpleName(), message));
-    }
-}
