--- src/main/java/com/jarqprog/artGallery/domain/exception/ExceptionInfo.java
@@ -1,11 +1,0 @@
-package com.jarqprog.artGallery.domain.exception;
-
-import java.time.LocalDateTime;
-
-public interface ExceptionInfo {
-
-    String getUuid();
-    long getHttpStatus();
-    LocalDateTime getDateTime();
-    String getMessage();
-}
