--- src/main/java/com/jarqprog/artGallery/domain/helper/ContactCreator.java
@@ -1,9 +1,0 @@
-package com.jarqprog.artGallery.domain.helper;
-
-import com.jarqprog.artGallery.domain.dto.ContactDTO;
-import com.jarqprog.artGallery.domain.dto.useCaseDTO.RegistrationDTO;
-
-public interface ContactCreator {
-
-    ContactDTO createContactFromRegistrationDTO(RegistrationDTO registrationDTO);
-}
