--- src/main/java/com/jarqprog/artGallery/domain/helper/PasswordValidator.java
@@ -1,7 +1,0 @@
-package com.jarqprog.artGallery.domain.helper;
-
-public interface PasswordValidator {
-
-    boolean validate(final String password);
-
-}
