--- src/main/java/com/jarqprog/artGallery/domain/helper/UserCreator.java
@@ -1,12 +1,0 @@
-package com.jarqprog.artGallery.domain.helper;
-
-
-import com.jarqprog.artGallery.domain.dto.ContactDTO;
-import com.jarqprog.artGallery.domain.dto.UserDTO;
-import com.jarqprog.artGallery.domain.dto.useCaseDTO.RegistrationDTO;
-
-public interface UserCreator {
-
-    UserDTO createUserFromRegistrationDTO(RegistrationDTO registrationDTO, ContactDTO contactDTO);
-
-}
