--- src/main/java/com/jarqprog/artGallery/domain/helper/UserRegistrationValidator.java
@@ -1,10 +1,0 @@
-package com.jarqprog.artGallery.domain.helper;
-
-import com.jarqprog.artGallery.domain.dto.useCaseDTO.RegistrationDTO;
-import com.jarqprog.artGallery.domain.exception.CreationException;
-
-public interface UserRegistrationValidator {
-
-    void validateContactData(RegistrationDTO registrationDTO) throws CreationException;
-    void validateUserData(RegistrationDTO registrationDTO) throws CreationException;
-}
