--- src/main/java/com/jarqprog/artGallery/springData/PersistenceConfig.java
@@ -1,66 +1,0 @@
-package com.jarqprog.artGallery.springData;
-
-import com.jarqprog.artGallery.springData.databaseConfig.DatabaseConfig;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.ComponentScan;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.dao.annotation.PersistenceExceptionTranslationPostProcessor;
-import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
-import org.springframework.jdbc.datasource.DriverManagerDataSource;
-import org.springframework.orm.jpa.JpaTransactionManager;
-import org.springframework.orm.jpa.JpaVendorAdapter;
-import org.springframework.orm.jpa.LocalContainerEntityManagerFactoryBean;
-import org.springframework.orm.jpa.vendor.HibernateJpaVendorAdapter;
-import org.springframework.transaction.PlatformTransactionManager;
-import org.springframework.transaction.annotation.EnableTransactionManagement;
-
-import javax.persistence.EntityManagerFactory;
-import javax.sql.DataSource;
-
-@Configuration
-@EnableTransactionManagement
-@EnableJpaRepositories(basePackages = "com.jarqprog.artGallery.springData.repository" )
-@ComponentScan({"com.jarqprog.artGallery.springData.databaseConfig",
-        "com.jarqprog.artGallery.domain.helper.implementation"})
-public class PersistenceConfig {
-
-    @Bean
-    protected DataSource dataSource(@Autowired DatabaseConfig databaseConfig) {
-        assert databaseConfig != null;
-        DriverManagerDataSource dataSource = new DriverManagerDataSource();
-        dataSource.setDriverClassName(databaseConfig.getDriverClass());
-        dataSource.setUrl(databaseConfig.getUrl());
-        dataSource.setUsername(databaseConfig.getUser());
-        dataSource.setPassword(databaseConfig.getPassword());
-        return dataSource;
-    }
-
-    @Bean
-    public LocalContainerEntityManagerFactoryBean entityManagerFactory(@Autowired DataSource dataSource,
-                                                                       @Autowired DatabaseConfig databaseConfig) {
-        assert dataSource != null;
-        assert databaseConfig != null;
-        LocalContainerEntityManagerFactoryBean em
-                = new LocalContainerEntityManagerFactoryBean();
-        em.setDataSource(dataSource);
-        em.setPackagesToScan("com.jarqprog.artGallery.domain.entity");
-
-        JpaVendorAdapter vendorAdapter = new HibernateJpaVendorAdapter();
-        em.setJpaVendorAdapter(vendorAdapter);
-        em.setJpaProperties(databaseConfig.getJPAProperties());
-        return em;
-    }
-
-    @Bean
-    public PlatformTransactionManager transactionManager(EntityManagerFactory emf) {
-        JpaTransactionManager transactionManager = new JpaTransactionManager();
-        transactionManager.setEntityManagerFactory(emf);
-        return transactionManager;
-    }
-
-    @Bean
-    public PersistenceExceptionTranslationPostProcessor exceptionTranslation(){
-        return new PersistenceExceptionTranslationPostProcessor();
-    }
-}
