+++ src/main/java/com/jarqprog/artGallery/springData/components/ContactCreator.java
@@ -1,0 +1,9 @@
+package com.jarqprog.artGallery.springData.components;
+
+import com.jarqprog.artGallery.domain.dto.ContactDTO;
+import com.jarqprog.artGallery.domain.dto.useCaseDTO.RegistrationDTO;
+
+public interface ContactCreator {
+
+    ContactDTO createContactFromRegistrationDTO(RegistrationDTO registrationDTO);
+}
