+++ src/main/java/com/jarqprog/artGallery/springData/components/PasswordValidator.java
@@ -1,0 +1,7 @@
+package com.jarqprog.artGallery.springData.components;
+
+public interface PasswordValidator {
+
+    boolean validate(final String password);
+
+}
