+++ src/main/java/com/jarqprog/artGallery/springData/config/JpaAuditingConfig.java
@@ -1,0 +1,23 @@
+package com.jarqprog.artGallery.springData.config;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.data.domain.AuditorAware;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
+
+import java.util.Optional;
+
+@Configuration
+@EnableJpaAuditing(auditorAwareRef = "auditorProvider")
+public class JpaAuditingConfig {
+
+    @Bean
+    public AuditorAware<String> auditorProvider() {
+        //todo to check
+        /*
+          if you are using spring springSecurity, you can get the currently logged username with following code segment.
+          SecurityContextHolder.getContext().getAuthentication().getName()
+         */
+        return () -> Optional.ofNullable("auditorProvider");
+    }
+}
