+++ src/main/java/com/jarqprog/artGallery/springData/exceptions/ExceptionInfo.java
@@ -1,0 +1,11 @@
+package com.jarqprog.artGallery.springData.exceptions;
+
+import java.time.LocalDateTime;
+
+public interface ExceptionInfo {
+
+    String getUuid();
+    long getHttpStatus();
+    LocalDateTime getDateTime();
+    String getMessage();
+}
