+++ src/main/java/com/jarqprog/artGallery/springWebMVC/service/WebRegistrationService.java
@@ -3,8 +3,8 @@ package com.jarqprog.artGallery.springWebMVC.service;
 import com.jarqprog.artGallery.domain.dto.ContactDTO;
 import com.jarqprog.artGallery.domain.dto.UserDTO;
 import com.jarqprog.artGallery.domain.dto.useCaseDTO.RegistrationDTO;
-import com.jarqprog.artGallery.domain.helper.ContactCreator;
-import com.jarqprog.artGallery.domain.helper.UserCreator;
+import com.jarqprog.artGallery.springData.components.ContactCreator;
+import com.jarqprog.artGallery.springData.components.UserCreator;
 import com.jarqprog.artGallery.domain.useCases.RegistrationService;
 
 import org.springframework.beans.factory.annotation.Autowired;
